genai-deployer/backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
jinja2==3.1.4
pyyaml==6.0.2
kubernetes==30.1.0
requests==2.32.3
